fix remove project problem
while deleting projects, the main component where the add task btn exists wasn't being rendered. This is because the whole p tag had the onclick event handler and the tabid got set to the title. now the tabid get sets to inbox for which after deleting a project the inbox tab will be shown